fix: missing nativescript and metadata validation
