+++ packages/core/src/shared-wallets/add-shared-wallet/restore-flow/types.ts
@@ -1,14 +1,5 @@
 import { CoSigner, QuorumOptionValue } from '../creation-flow';
 
-export enum FileErrorMessage {
-  INVALID_KEY = 'Invalid key',
-  UNRECOGNIZED = 'File is unrecognized',
-}
-
-export interface FileValidationError extends Error {
-  message: FileErrorMessage;
-}
-
 export enum SharedWalletRestorationStep {
   Done = 'Done',
   Import = 'Import',
